Experimental ABM batch execution
The theory is that there is significant cost to context switching
between the C++ engine and Lua mod code to call the ABM action
function for EACH node affected.  This change batches up pos/node
pairs and makes one big call to Lua, passing in the entire batch
as an array.

Because ABM executions can change the active object counts, and
nodes within the affected area (which would change ABM eligibility)
this new "batched" mode is not compatible with existing ABMs and
thus needs to be requested separately.  Lua code in builtin
provides a similar per-node action interface by looping through
the batch and dispatching per-node entirely on the Lua side.

N.B. the current implementation probably suffers memory leaks.

In my experiments, I was not able to measure any difference in
overall system performance, neither in step size, nor in
number of nodes processed per step, even though I confirmed that
I'm receiving at least some batches with significant numbers of
nodes in them, and the Lua profiler indicates that ABMs are
already well-optimized, and changing the number of ABMs had the
biggest performance impact on the game.